requirements.txt
PyYAML
aiohttp
aioredis==1.3.1
sentry-sdk==0.15.1
psutil
prometheus_async[aiohttp]